% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostdata-submitted.R
\name{check_hostdata_submitted}
\alias{check_hostdata_submitted}
\title{Check if deposit is submitted}
\usage{
check_hostdata_submitted(hostdata)
}
\arguments{
\item{hostdata}{data returned from service API}
}
\value{
logical. TRUE if the deposit has been published to zenodo
}
\description{
Checks to see if a deposit is using zenodo and has been submitted according to its hostdata
}
\details{
See \url{https://developers.zenodo.org/#representation} for more details about the submitted field.
}
