% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostdata-submitted.R
\name{correct_download_urls}
\alias{correct_download_urls}
\title{Correct download urls for submitted deposits}
\usage{
correct_download_urls(hostdata)
}
\arguments{
\item{hostdata}{data returned from service API}
}
\value{
returns self$hostdata.
}
\description{
Removes "draft/" from the URL of deposits that have been published on zenodo.
}
