% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deposits-package.R
\docType{package}
\name{deposits-package}
\alias{deposits}
\alias{deposits-package}
\title{deposits: A universal client for depositing and accessing research data anywhere.}
\description{
A universal client for depositing and accessing research data anywhere. Currently supported services are zenodo and figshare.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/deposits/}
  \item Report bugs at \url{https://github.com/ropenscilabs/deposits/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
