% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-template.R
\name{deposits_meta_from_file}
\alias{deposits_meta_from_file}
\title{Read a metadata \code{JSON} file and convert to a metadata object.}
\usage{
deposits_meta_from_file(filename = NULL, id = "my-id")
}
\arguments{
\item{filename}{Name of completed metadata file generated from
\link{deposits_metadata_template}.}

\item{id}{A unique identifier for the deposit, as a single character string.}
}
\description{
Metadata templates can be generated with \link{deposits_metadata_template},
and values manually entered. This functions loads the completed template and
converts it to a metadata object.
}
\examples{
filename <- tempfile (fileext = ".json")
deposits_metadata_template (filename)
m <- readLines (filename)
m [grep ("Title", m)] <- "  \"Title\": \"New Title\""
m [grep ("Type", m)] <- "  \"Type\": \"Software\""
m [grep ("Description", m)] <-
    "  \"Description\": \"Description of software\""
# Values can be entered in JSON format:
m [grep ("TableOfContents", m)] <-
    "  \"TableOfContents\": {\"one\": \"First\", \"two\": \"Second\"}"
dc <- deposits_meta_from_file (filename, id = "my-id")
}
\seealso{
Other meta: 
\code{\link{dcmi_terms}()},
\code{\link{deposits_metadata_template}()}
}
\concept{meta}
