% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{deposits_services}
\alias{deposits_services}
\title{List all deposits services and associated URLs}
\usage{
deposits_services()
}
\value{
A \code{data.frame} with \code{name} and \code{url} values for each accessible
service.
}
\description{
List all deposits services and associated URLs
}
\examples{
s <- deposits_services ()
}
\seealso{
Other misc: 
\code{\link{deposits_metadata_template}()}
}
\concept{misc}
