% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{depositsClient}
\alias{depositsClient}
\title{depositsClient}
\value{
a \code{depositsClient} class (R6 class)
}
\description{
R6 class for constructing deposits queries
}
\examples{
\dontrun{
# make a client
d <- depositsClient$new ("zenodo") # or:
d <- depositsClient$new ("figshare")

# methods
d$list_deposits ()
}
}
\concept{client}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{(character) of deposits server}

\item{\code{url}}{(character) list of fragments}

\item{\code{headers}}{list of named headers}

\item{\code{schema}}{holds schema}

\item{\code{result}}{holds result from http request}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{depositsClient$new()}}
\item \href{#method-print}{\code{depositsClient$print()}}
\item \href{#method-ping}{\code{depositsClient$ping()}}
\item \href{#method-list_deposits}{\code{depositsClient$list_deposits()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{depositsClient} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$new(name, headers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(character) of a deposits service (see
\link{deposits_services}).}

\item{\code{headers}}{Any acceptable headers. See examples}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{depositsClient} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print method for the \code{depositsClient} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ping"></a>}}
\if{latex}{\out{\hypertarget{method-ping}{}}}
\subsection{Method \code{ping()}}{
ping a deposits server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$ping(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{curl options passed on to \link[crul:verb-HEAD]{crul::verb-HEAD}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if successful response, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_deposits"></a>}}
\if{latex}{\out{\hypertarget{method-list_deposits}{}}}
\subsection{Method \code{list_deposits()}}{
List own deposits for given service
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$list_deposits(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{curl options passed on to \link[crul:verb-HEAD]{crul::verb-HEAD}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of deposits.
}
}
}
