% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figshare-categories.R
\name{figshare_categories}
\alias{figshare_categories}
\title{Select figshare categories and return corresponding integer identifier.}
\usage{
figshare_categories()
}
\description{
These identifiers should then be added in deposit metadata as, for example,
\code{subject(categories=c(1,2))}.
}
\seealso{
Other meta: 
\code{\link{dcmi_terms}()},
\code{\link{deposits_metadata_template}()}
}
\concept{meta}
