% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deposits-package.R
\docType{package}
\name{deposits-package}
\alias{deposits}
\alias{deposits-package}
\title{deposits: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
