% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{depositsClient}
\alias{depositsClient}
\title{depositsClient}
\value{
a \code{depositsClient} class (R6 class)
}
\description{
R6 class for constructing deposits queries
}
\examples{
x <- depositsClient$new()
x

\dontrun{
# make a client
token <- Sys.getenv("ZENODO_TOKEN")
cli <- depositsClient$new(
  url = "https://zenodo.org/api",
  headers = list(Authorization = paste0("Bearer ", token))
)

# methods
## ping - hopefully you get TRUE
cli$ping()

}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{url}}{(character) list of fragments}

\item{\code{headers}}{list of named headers}

\item{\code{schema}}{holds schema}

\item{\code{result}}{holds result from http request}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{depositsClient$new()}}
\item \href{#method-print}{\code{depositsClient$print()}}
\item \href{#method-ping}{\code{depositsClient$ping()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{depositsClient} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$new(url, headers)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{(character) URL for the deposits schema}

\item{\code{headers}}{Any acceptable headers, a named list. See examples}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{depositsClient} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print method for the \code{depositsClient} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ping"></a>}}
\if{latex}{\out{\hypertarget{method-ping}{}}}
\subsection{Method \code{ping()}}{
ping a deposits server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$ping(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{curl options passed on to \link[crul:verb-HEAD]{crul::verb-HEAD}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if successful response, \code{FALSE} otherwise
}
}
}
