% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{depositsClient}
\alias{depositsClient}
\title{depositsClient}
\value{
a \code{depositsClient} class (R6 class)
}
\description{
R6 class for constructing deposits queries
}
\note{
Client should already contain metadata updated with the
'fill_metadata()' function.
}
\examples{
\dontrun{
# make a client
d <- depositsClient$new ("zenodo") # or:
d <- depositsClient$new ("figshare")

# methods
d$list_deposits ()

# Construct new deposit:
meta <- atom4R::DCEntry$new ()
meta$addDCTitle ("New deposit")
cli <- depositsClient$new ("zenodo", sandbox = TRUE, meta = meta)
res <- cli$new_deposit () # upload metadata to construct new zenodo deposit
deposit_id <- res$id # or "entity_id" for "figshare"
cli$retrieve_deposit (deposit_id = deposit_id)

# Update deposit:
# Note that updating atom4R::DCentry metadata objects generally requires
# deleting old element and replacing with new one.
meta$delDCTitle (meta$title [[1]])
meta$addDCTitle ("This is a new title")
cli$fill_metadata (meta)
cli$update_deposit (deposit_id = deposit_id)

# Delete deposit:
cli$delete_deposit (deposit_id = deposit_id)
}
}
\concept{client}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{(character) of deposits server}

\item{\code{sandbox}}{Connect client with sandbox if \code{TRUE} (zenodo only)}

\item{\code{url}}{(character) list of fragments}

\item{\code{headers}}{list of named headers}

\item{\code{schema}}{holds schema}

\item{\code{result}}{holds result from http request}

\item{\code{metadata}}{holds metadata}

\item{\code{term_map}}{Map between DCMI and deposit terms}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-depositsClient-new}{\code{depositsClient$new()}}
\item \href{#method-depositsClient-print}{\code{depositsClient$print()}}
\item \href{#method-depositsClient-ping}{\code{depositsClient$ping()}}
\item \href{#method-depositsClient-list_deposits}{\code{depositsClient$list_deposits()}}
\item \href{#method-depositsClient-delete_deposit}{\code{depositsClient$delete_deposit()}}
\item \href{#method-depositsClient-fill_metadata}{\code{depositsClient$fill_metadata()}}
\item \href{#method-depositsClient-new_deposit}{\code{depositsClient$new_deposit()}}
\item \href{#method-depositsClient-update_deposit}{\code{depositsClient$update_deposit()}}
\item \href{#method-depositsClient-upload_file}{\code{depositsClient$upload_file()}}
\item \href{#method-depositsClient-retrieve_deposit}{\code{depositsClient$retrieve_deposit()}}
\item \href{#method-depositsClient-download_file}{\code{depositsClient$download_file()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{depositsClient} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$new(name, metadata = NULL, sandbox = FALSE, headers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(character) of a deposits service (see
\link{deposits_services}).}

\item{\code{metadata}}{Either name (or full path) or a local file containing
metadata constructed with \link{deposits_metadata_template}, or an
\pkg{atom4R} \code{DCEntry} object containing metadata, either constructed
directly via \pkg{atom4R} routines, or via
\link{deposits_meta_to_dcmi}.}

\item{\code{sandbox}}{If \code{TRUE}, connect client to sandbox, rather than
actual API endpoint (for "zenodo" only).}

\item{\code{headers}}{Any acceptable headers. See examples in \pkg{crul}
package.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{depositsClient} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-print"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-print}{}}}
\subsection{Method \code{print()}}{
print method for the \code{depositsClient} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-ping"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-ping}{}}}
\subsection{Method \code{ping()}}{
ping a deposits server to check authentication
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$ping()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if successful response, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-list_deposits"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-list_deposits}{}}}
\subsection{Method \code{list_deposits()}}{
List own deposits for given service
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$list_deposits()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of deposits.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-delete_deposit"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-delete_deposit}{}}}
\subsection{Method \code{delete_deposit()}}{
Deleted a nominated deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$delete_deposit(deposit_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{Integer identifer of deposit (generally obtained
from \code{list_deposits} method).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \pkg{crul} response object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-fill_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-fill_metadata}{}}}
\subsection{Method \code{fill_metadata()}}{
Fill deposits client with metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$fill_metadata(metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{An \pkg{atom4R} \code{DCEntry} object containing metadata,
either constructed directly via \pkg{atom4R} routines, or via
\link{deposits_meta_to_dcmi}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Modified form of the deposits client with metadata inserted.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-new_deposit"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-new_deposit}{}}}
\subsection{Method \code{new_deposit()}}{
Create a new deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$new_deposit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} with details of the newly created deposit.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-update_deposit"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-update_deposit}{}}}
\subsection{Method \code{update_deposit()}}{
Update metadata for specified deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$update_deposit(deposit_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{The 'id' number of deposit to update.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with details of newly updated deposit
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-upload_file"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-upload_file}{}}}
\subsection{Method \code{upload_file()}}{
Upload file to an existing deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$upload_file(deposit_id, path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{The 'id' number of deposit which file it to be
uploaded to. (generally obtained from \code{list_deposits} method).}

\item{\code{path}}{Path to local file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with details on newly uploaded file, including
upload and download URLs, and file details.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-retrieve_deposit"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-retrieve_deposit}{}}}
\subsection{Method \code{retrieve_deposit()}}{
Retrieve information on specified deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$retrieve_deposit(deposit_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{The 'id' number of deposit for which information is
to be retrieved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} containing full data of specified deposit.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-download_file"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-download_file}{}}}
\subsection{Method \code{download_file()}}{
Download a specified 'filename' from a deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$download_file(deposit_id, filename, path = NULL, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{The 'id' number of deposit for which information is
to be retrieved.}

\item{\code{filename}}{The name of the file to be download as specified in
the deposit.}

\item{\code{path}}{The local directory where file is to be downloaded.}

\item{\code{quiet}}{If \code{FALSE}, display download progress.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The full path of the downloaded file.
}
}
}
