% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-main.R
\name{depositsClient}
\alias{depositsClient}
\title{depositsClient}
\value{
a \code{depositsClient} class (R6 class)
}
\description{
R6 class for constructing deposits queries
}
\note{
Client should already contain metadata updated with the
'deposit_fill_metadata()' function.
}
\examples{
\dontrun{
# make a client
cli <- depositsClient$new ("zenodo") # or:
cli <- depositsClient$new ("figshare")
print (cli)

# methods
cli$deposits_list ()

# Fill depositsClient metadata
metadata <- list (
    title = "New Title",
    abstract = "This is the abstract",
    creator = list ("A. Person", "B. Person")
)
cli$deposit_fill_metadata (metadata)
print (cli)

# or pass metadata directly at construction of new client
cli <- depositsClient$new ("figshare", metadata = metadata)
}

## ------------------------------------------------
## Method `depositsClient$deposits_search`
## ------------------------------------------------

\dontrun{
cli <- depositsClient$new (service = "figshare")
search_results <- cli$deposits_search (
    search_string = "Text string query",
    page_size = 5L
)
}
}
\concept{client}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{service}}{(character) of deposits host service.}

\item{\code{sandbox}}{(logical) Connect client with sandbox if \code{TRUE}
(zenodo only)}

\item{\code{deposits}}{(data.frame) Current deposits hosted on service, one
row per deposit.}

\item{\code{url_base}}{(character) Base URL of host service API}

\item{\code{url_service}}{(character) URL of deposit service}

\item{\code{id}}{(integer) Deposit identifier from host service.}

\item{\code{headers}}{(list) list of named headers}

\item{\code{hostdata}}{(list) Data as stored by host platform}

\item{\code{metadata}}{holds list of DCMI-compliant metadata.}

\item{\code{metadata_service}}{holds metadata converted to specific format
requires by service. Derived from \code{metadata}.}

\item{\code{term_map}}{(data.frame) Map between DCMI and deposit terms for
specified host service.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-depositsClient-new}{\code{depositsClient$new()}}
\item \href{#method-depositsClient-print}{\code{depositsClient$print()}}
\item \href{#method-depositsClient-deposits_list}{\code{depositsClient$deposits_list()}}
\item \href{#method-depositsClient-deposits_search}{\code{depositsClient$deposits_search()}}
\item \href{#method-depositsClient-deposit_delete}{\code{depositsClient$deposit_delete()}}
\item \href{#method-depositsClient-deposit_fill_metadata}{\code{depositsClient$deposit_fill_metadata()}}
\item \href{#method-depositsClient-deposit_new}{\code{depositsClient$deposit_new()}}
\item \href{#method-depositsClient-deposit_service}{\code{depositsClient$deposit_service()}}
\item \href{#method-depositsClient-deposit_update}{\code{depositsClient$deposit_update()}}
\item \href{#method-depositsClient-deposit_upload_file}{\code{depositsClient$deposit_upload_file()}}
\item \href{#method-depositsClient-deposit_retrieve}{\code{depositsClient$deposit_retrieve()}}
\item \href{#method-depositsClient-deposit_download_file}{\code{depositsClient$deposit_download_file()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{depositsClient} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$new(service, metadata = NULL, sandbox = FALSE, headers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{service}}{(character) of a deposits service (see
\link{deposits_services}).}

\item{\code{metadata}}{Either of one two possible ways of defining
metadata:
\itemize{
\item The name (or full path) or a local file containing
metadata constructed with \link{deposits_metadata_template};
\item A names list of metadata with names matching values given by
\link{dcmi_terms}, and values specified as individual character
strings or lists for multiple entries.
}}

\item{\code{sandbox}}{If \code{TRUE}, connect client to sandbox, rather than
actual API endpoint (for "zenodo" only).}

\item{\code{headers}}{Any acceptable headers. See examples in \pkg{httr2}
package.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{depositsClient} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-print"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-print}{}}}
\subsection{Method \code{print()}}{
print method for the \code{depositsClient} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposits_list"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposits_list}{}}}
\subsection{Method \code{deposits_list()}}{
Update 'deposits' item of current deposits for given
service
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposits_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Updated 'deposits' client
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposits_search"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposits_search}{}}}
\subsection{Method \code{deposits_search()}}{
Search all public deposits
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposits_search(
  search_string = NULL,
  page_size = 10L,
  page_number = 1L,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_string}}{Single string to search for}

\item{\code{page_size}}{Number of records to return in one page}

\item{\code{page_number}}{Starting page for return results; used in
combination with 'page_size' for pagination.}

\item{\code{...}}{Named pairs of query parameters.
Zenodo parameters are described at
\url{https://developers.zenodo.org/#list36}, and currently include:
\itemize{
\item status: either "draft" or "published"
\item sort: either "bestmatch" (the default) or "mostrecent"
\item all_versions: Either "true" or "false"
\item communities: Search for deposits only within specified
communities
\item type: Return deposits only of specified type
\item subtype: Return deposits only of specified subtype
\item bound: A geolocation bounding box
\item custom: Custom keywords
}

Figshare parameters are described at
\url{https://docs.figshare.com/#articles_search}, and currently
include:
\itemize{
\item resource_doi: Only return deposits matching this 'resource_doi'
\item item_type: Return deopsits of specified type (as integer).
\item doi: Only return deposits matching this DOI
\item handle: Only return deposits matching this handle
\item project_id: Only return deposits from within specified project
\item order: Order for sorting results; one of "published_date",
"modified_date", "views", "shares", "downloads", or "cites"
\item search_for: Search term.
\item order_direction: "asc" or "desc"
\item institution: Only return deposits from specified institution
(as integer)
\item group: Only return deposits from specified group (as integer)
\item published_since: Only return deposits published since specified
date (as YYYY-MM-DD)
\item modified_since: Only return deposits modified since specified
date (as YYYY-MM-DD)
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} of data on deposits matching search parameters
(with format depending on the deposits service.)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
cli <- depositsClient$new (service = "figshare")
search_results <- cli$deposits_search (
    search_string = "Text string query",
    page_size = 5L
)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_delete"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_delete}{}}}
\subsection{Method \code{deposit_delete()}}{
Deleted a nominated deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_delete(deposit_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{Integer identifier of deposit (generally obtained
from \code{list_deposits} method).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated 'deposits' client
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_fill_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_fill_metadata}{}}}
\subsection{Method \code{deposit_fill_metadata()}}{
Fill deposits client with metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_fill_metadata(metadata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Either one of two possible ways of defining
metadata:
\itemize{
\item The name (or full path) or a local file containing
metadata constructed with \link{deposits_metadata_template};
\item A names list of metadata with names matching values given by
\link{dcmi_terms}, and values specified as individual character
strings or lists for multiple entries.
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated deposits client with metadata inserted.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_new"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_new}{}}}
\subsection{Method \code{deposit_new()}}{
Create a new deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_new(quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quiet}}{If \code{FALSE} (default), print integer identifier of newly
created deposit.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated deposits client which includes data on new deposit
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_service"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_service}{}}}
\subsection{Method \code{deposit_service()}}{
Create a new \code{depositsClient} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_service(service = NULL, sandbox = FALSE, headers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{service}}{(character) of a deposits service (see
\link{deposits_services}).}

\item{\code{sandbox}}{If \code{TRUE}, connect client to sandbox, rather than
actual API endpoint (for "zenodo" only).}

\item{\code{headers}}{Any acceptable headers. See examples in \pkg{httr2}
package.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_update"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_update}{}}}
\subsection{Method \code{deposit_update()}}{
Update metadata for specified deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_update(deposit_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{The 'id' number of deposit to update. If not
specified, the 'id' value of current deposits client is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_upload_file"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_upload_file}{}}}
\subsection{Method \code{deposit_upload_file()}}{
Upload file to an existing deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_upload_file(deposit_id = NULL, path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{The 'id' number of deposit which file is to be
uploaded to. If not specified, the 'id' value of current deposits
client is used.}

\item{\code{path}}{Path to local file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated 'deposits' client
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_retrieve}{}}}
\subsection{Method \code{deposit_retrieve()}}{
Retrieve information on specified deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_retrieve(deposit_id, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{The 'id' number of deposit for which information is
to be retrieved.}

\item{\code{quiet}}{If \code{FALSE} (default), display information on screen on
any issues encountered in retrieving deposit.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated 'deposits' client
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_download_file"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_download_file}{}}}
\subsection{Method \code{deposit_download_file()}}{
Download a specified 'filename' from a deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_download_file(
  deposit_id = NULL,
  filename,
  path = NULL,
  overwrite = FALSE,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{The 'id' number of deposit which file is to be
downloaded from. If not specified, the 'id' value of current deposits
client is used.}

\item{\code{filename}}{The name of the file to be download as specified in
the deposit.}

\item{\code{path}}{The local directory where file is to be downloaded.}

\item{\code{overwrite}}{Do not overwrite existing files unless set to
\code{TRUE}.}

\item{\code{quiet}}{If \code{FALSE}, display download progress.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The full path of the downloaded file.
}
}
}
