% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-main.R
\name{depositsClient}
\alias{depositsClient}
\title{depositsClient}
\value{
a \code{depositsClient} class (R6 class)
}
\description{
R6 class for constructing deposits queries
}
\note{
Client should already contain metadata updated with the
'deposit_fill_metadata()' function.
}
\examples{
\dontrun{
# make a client
cli <- depositsClient$new ("zenodo") # or:
cli <- depositsClient$new ("figshare")
print (cli)

# methods
cli$deposits_list ()

# Fill depositsClient metadata
metadata <- list (
    title = "New Title",
    abstract = "This is the abstract",
    creator = list ("A. Person", "B. Person")
)
cli$deposit_fill_metadata (metadata)
print (cli)

# or pass metadata directly at construction of new client
cli <- depositsClient$new ("figshare", metadata = metadata)
}
}
\concept{client}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{service}}{(character) of deposits host service.}

\item{\code{sandbox}}{(logical) Connect client with sandbox if \code{TRUE}
(zenodo only)}

\item{\code{deposits}}{(data.frame) Current deposits hosted on service, one
row per deposit.}

\item{\code{url_base}}{(character) Base URL of host service API}

\item{\code{url_deposit}}{(character) URL of deposit.}

\item{\code{id}}{(integer) Deposit identifier from host service.}

\item{\code{headers}}{(list) list of named headers}

\item{\code{hostdata}}{(list) Data as stored by host platform}

\item{\code{metadata}}{(\code{atom4R::DCEntry}) holds metadata}

\item{\code{term_map}}{(data.frame) Map between DCMI and deposit terms for
specified host service.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-depositsClient-new}{\code{depositsClient$new()}}
\item \href{#method-depositsClient-print}{\code{depositsClient$print()}}
\item \href{#method-depositsClient-deposits_list}{\code{depositsClient$deposits_list()}}
\item \href{#method-depositsClient-deposit_delete}{\code{depositsClient$deposit_delete()}}
\item \href{#method-depositsClient-deposit_fill_metadata}{\code{depositsClient$deposit_fill_metadata()}}
\item \href{#method-depositsClient-deposit_new}{\code{depositsClient$deposit_new()}}
\item \href{#method-depositsClient-deposit_update}{\code{depositsClient$deposit_update()}}
\item \href{#method-depositsClient-deposit_upload_file}{\code{depositsClient$deposit_upload_file()}}
\item \href{#method-depositsClient-deposit_retrieve}{\code{depositsClient$deposit_retrieve()}}
\item \href{#method-depositsClient-deposit_download_file}{\code{depositsClient$deposit_download_file()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{depositsClient} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$new(service, metadata = NULL, sandbox = FALSE, headers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{service}}{(character) of a deposits service (see
\link{deposits_services}).}

\item{\code{metadata}}{Either of one three possible ways of defining
metadata:
\itemize{
\item The name (or full path) or a local file containing
metadata constructed with \link{deposits_metadata_template};
\item A names list of metadata with names matching values given by
\link{dcmi_terms}, and values specified as individual character
strings or lists for multiple entries.
\item An \pkg{atom4R} \code{DCEntry} object containing metadata, either
constructed directly via \pkg{atom4R} routines, or via
\link{deposits_meta_to_dcmi}.
}}

\item{\code{sandbox}}{If \code{TRUE}, connect client to sandbox, rather than
actual API endpoint (for "zenodo" only).}

\item{\code{headers}}{Any acceptable headers. See examples in \pkg{httr2}
package.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{depositsClient} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-print"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-print}{}}}
\subsection{Method \code{print()}}{
print method for the \code{depositsClient} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposits_list"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposits_list}{}}}
\subsection{Method \code{deposits_list()}}{
Update 'deposits' item of current deposits for given
service
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposits_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Updated 'deposits' client
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_delete"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_delete}{}}}
\subsection{Method \code{deposit_delete()}}{
Deleted a nominated deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_delete(deposit_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{Integer identifier of deposit (generally obtained
from \code{list_deposits} method).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} is deposit successfully deleted, otherwise \code{FALSE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_fill_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_fill_metadata}{}}}
\subsection{Method \code{deposit_fill_metadata()}}{
Fill deposits client with metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_fill_metadata(metadata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Either of one three possible ways of defining
metadata:
\itemize{
\item The name (or full path) or a local file containing
metadata constructed with \link{deposits_metadata_template};
\item A names list of metadata with names matching values given by
\link{dcmi_terms}, and values specified as individual character
strings or lists for multiple entries.
\item An \pkg{atom4R} \code{DCEntry} object containing metadata, either
constructed directly via \pkg{atom4R} routines, or via
\link{deposits_meta_to_dcmi}.
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Modified form of the deposits client with metadata inserted.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_new"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_new}{}}}
\subsection{Method \code{deposit_new()}}{
Create a new deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} with details of the newly created deposit.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_update"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_update}{}}}
\subsection{Method \code{deposit_update()}}{
Update metadata for specified deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_update(deposit_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{The 'id' number of deposit to update. If not
specified, the 'id' value of current deposits client is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_upload_file"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_upload_file}{}}}
\subsection{Method \code{deposit_upload_file()}}{
Upload file to an existing deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_upload_file(deposit_id = NULL, path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{The 'id' number of deposit which file is to be
uploaded to. If not specified, the 'id' value of current deposits
client is used.}

\item{\code{path}}{Path to local file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with details on newly uploaded file, including
upload and download URLs, and file details.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_retrieve}{}}}
\subsection{Method \code{deposit_retrieve()}}{
Retrieve information on specified deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_retrieve(deposit_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{The 'id' number of deposit for which information is
to be retrieved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} containing full data of specified deposit.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositsClient-deposit_download_file"></a>}}
\if{latex}{\out{\hypertarget{method-depositsClient-deposit_download_file}{}}}
\subsection{Method \code{deposit_download_file()}}{
Download a specified 'filename' from a deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$deposit_download_file(
  deposit_id = NULL,
  filename,
  path = NULL,
  overwrite = FALSE,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deposit_id}}{The 'id' number of deposit which file is to be
downloaded from. If not specified, the 'id' value of current deposits
client is used.}

\item{\code{filename}}{The name of the file to be download as specified in
the deposit.}

\item{\code{path}}{The local directory where file is to be downloaded.}

\item{\code{overwrite}}{Do not overwrite existing files unless set to
\code{TRUE}.}

\item{\code{quiet}}{If \code{FALSE}, display download progress.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The full path of the downloaded file.
}
}
}
