% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{depositsClient}
\alias{depositsClient}
\title{depositsClient}
\value{
a \code{depositsClient} class (R6 class)
}
\description{
R6 class for constructing deposits queries
}
\examples{
\dontrun{
# make a client
d <- depositsClient$new ("zenodo") # or:
d <- depositsClient$new ("figshare")

# methods
d$list_deposits ()
}
}
\concept{client}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{(character) of deposits server}

\item{\code{sandbox}}{Connect client with sandbox if \code{TRUE} (zenodo only)}

\item{\code{url}}{(character) list of fragments}

\item{\code{headers}}{list of named headers}

\item{\code{schema}}{holds schema}

\item{\code{result}}{holds result from http request}

\item{\code{metadata}}{holds metadata}

\item{\code{term_map}}{Map between DCMI and deposit terms}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{depositsClient$new()}}
\item \href{#method-print}{\code{depositsClient$print()}}
\item \href{#method-ping}{\code{depositsClient$ping()}}
\item \href{#method-list_deposits}{\code{depositsClient$list_deposits()}}
\item \href{#method-delete_deposit}{\code{depositsClient$delete_deposit()}}
\item \href{#method-fill_metadata}{\code{depositsClient$fill_metadata()}}
\item \href{#method-new_deposit}{\code{depositsClient$new_deposit()}}
\item \href{#method-upload_file}{\code{depositsClient$upload_file()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{depositsClient} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$new(name, metadata = NULL, sandbox = FALSE, headers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(character) of a deposits service (see
\link{deposits_services}).}

\item{\code{metadata}}{An \pkg{atom4R} \code{DCEntry} object containing metadata,
either constructed directly via \pkg{atom4R} routines, or via}

\item{\code{sandbox}}{If \code{TRUE}, connect client to sandbox, rather than
actual API endpoint (for "zenodo" only).}

\item{\code{headers}}{Any acceptable headers. See examples}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{depositsClient} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print method for the \code{depositsClient} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ping"></a>}}
\if{latex}{\out{\hypertarget{method-ping}{}}}
\subsection{Method \code{ping()}}{
ping a deposits server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$ping(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{curl options passed on to \link[crul:verb-HEAD]{crul::verb-HEAD}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if successful response, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_deposits"></a>}}
\if{latex}{\out{\hypertarget{method-list_deposits}{}}}
\subsection{Method \code{list_deposits()}}{
List own deposits for given service
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$list_deposits(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{curl options passed on to \link[crul:verb-HEAD]{crul::verb-HEAD}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of deposits.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_deposit"></a>}}
\if{latex}{\out{\hypertarget{method-delete_deposit}{}}}
\subsection{Method \code{delete_deposit()}}{
Deleted a nominated deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$delete_deposit(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Integer identifer of deposit (generally from
\link{list_deposits}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \pkg{crul} response object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-fill_metadata}{}}}
\subsection{Method \code{fill_metadata()}}{
Fill deposits client with metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$fill_metadata(metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{An \pkg{atom4R} \code{DCEntry} object containing metadata,
either constructed directly via \pkg{atom4R} routines, or via
\link{deposits_meta_to_dcmi}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Modified form of the deposits client with metadata inserted.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new_deposit"></a>}}
\if{latex}{\out{\hypertarget{method-new_deposit}{}}}
\subsection{Method \code{new_deposit()}}{
Create a new deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$new_deposit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \pkg{crul} response object. Results can be extracted with
\code{jsonlite::fromJSON(result$parse(format="UTF-8"))}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_file"></a>}}
\if{latex}{\out{\hypertarget{method-upload_file}{}}}
\subsection{Method \code{upload_file()}}{
Upload file to an existing deposit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{depositsClient$upload_file(depost_id, path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depost_id}}{The 'id' number of deposit which file it to be uploaded to.}

\item{\code{path}}{Path to local file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \pkg{crul} response object containing full data of deposit.
including of uploaded file.
}
}
}
