% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{get_deposits_token}
\alias{get_deposits_token}
\title{Retrieve a token for a specified deposits service.}
\usage{
get_deposits_token(service = NULL)
}
\arguments{
\item{service}{Name of desired service; must be a value in the "name" column
of \link{deposits_services}.}
}
\value{
API token for nominated service.
}
\description{
Tokens should be stored as local environment variables, optionally defined in
a \verb{~/.Renviron} file, and should contain the name of the desired deposits
service.
}
