% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{get_deposits_token}
\alias{get_deposits_token}
\title{Retrieve a token for a specified deposits service.}
\usage{
get_deposits_token(service = NULL, sandbox = FALSE)
}
\arguments{
\item{service}{Name of desired service; must be a value in the "name" column
of \link{deposits_services}.}

\item{sandbox}{If \code{TRUE}, retrieve token for sandbox, rather than actual API.}
}
\value{
API token for nominated service.
}
\description{
Tokens should be stored as local environment variables, optionally defined in
a \verb{~/.Renviron} file. Tokens can be defined by using the pattern
\code{SERVICE_TOKEN}, where \code{SERVICE} is the name of the service in capital letters
followed by \verb{_SANDBOX} if the sandbox from e.g. zotero should be used. An example
would be \code{ZOTERO_SANDBOX_TOKEN} for the sandbox of zotero.
If these are not found, an heuristic algorythm is used to find the token name
which should contain the name of the desired deposits service.
}
\examples{
\dontrun{
token <- get_deposits_token (service = "figshare")
}
}
\concept{auth}
